Passing data from child to parent
TimerForm was converted from a functional component to a class component
to handle its own state since it is a form. Also connected TimerForm to
pass data up the hierarchy to NewTimerToggler then root App component
to handle the creation of a new timer, which modifies the state of
App (an array of timers).